% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{plink_qc}
\alias{plink_qc}
\title{Run the PLINK quality control for MonsterLM
\code{plink} is the path to the plink executable.
\code{out_dir} is the output directory.
\code{get_genotype} is a function that takes a chromosome number (\code{chr}) and returns the genotype file.
\code{get_allele} is a function that takes a chromosome number (\code{chr}) and returns the allele file.
\code{maf} is the minor allele frequency threshold.}
\usage{
plink_qc(plink, out_dir, get_genotype, get_allele, maf)
}
\description{
Run the PLINK quality control for MonsterLM
\code{plink} is the path to the plink executable.
\code{out_dir} is the output directory.
\code{get_genotype} is a function that takes a chromosome number (\code{chr}) and returns the genotype file.
\code{get_allele} is a function that takes a chromosome number (\code{chr}) and returns the allele file.
\code{maf} is the minor allele frequency threshold.
}
